Fixed use of Pusher by multiple threads
We noticed this because threads would get stuck accessing the
headers dictionary in the HttpClient class used by Pusher.
This often means that the dictionary has been corrupted by
concurrent access. Our application uses a singlton Pusher instance
which is shared by multiple threads. We're working around it in
our application by using an instance per thread instead of a
shared instance, but since this is a common patter maybe we should
make the client thread-safe?